requirements.txt
Flask==3.0.3
pydicom==2.4.4
pillow==10.4.0
numpy==2.1.1
requests==2.32.3
gunicorn==23.0.0
